+++ apps/docs/astro.config.mjs
@@ -1,14 +1,19 @@
 import react from "@astrojs/react";
 import { defineConfig } from "astro/config";
 
-import tailwind from "@astrojs/tailwind";
-
 import mdx from "@astrojs/mdx";
+
+import tailwindcss from "@tailwindcss/vite";
 
 // https://astro.build/config
 export default defineConfig({
-	integrations: [react(), tailwind(), mdx()],
+	integrations: [react(), mdx()],
+
 	devToolbar: {
 		enabled: false,
+	},
+
+	vite: {
+		plugins: [tailwindcss()],
 	},
 });
