+++ apps/docs/package.json
@@ -13,22 +13,22 @@
 	},
 	"dependencies": {
 		"@astrojs/check": "^0.9.4",
-		"@astrojs/mdx": "^3.1.9",
-		"@astrojs/react": "^3.6.3",
-		"@astrojs/tailwind": "^5.1.5",
+		"@astrojs/mdx": "^4.0.8",
+		"@astrojs/react": "^4.2.0",
 		"@bibleio/design": "workspace:*",
 		"@fontsource/jetbrains-mono": "^5.1.2",
 		"@fontsource/lora": "^5.1.1",
+		"@tailwindcss/vite": "^4.0.6",
 		"@types/react": "^18.3.18",
 		"@types/react-dom": "^18.3.5",
-		"astro": "^4.16.18",
+		"astro": "^5.3.0",
 		"gsap": "^3.12.7",
 		"prettier": "^3.5.1",
 		"prettier-plugin-astro": "^0.14.1",
 		"prettier-plugin-tailwindcss": "^0.6.11",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
-		"tailwindcss": "^3.4.17",
+		"tailwindcss": "^4.0.6",
 		"typescript": "^5.7.3"
 	}
 }
