--- apps/docs/src/env.d.ts
@@ -1,1 +1,0 @@
-/// <reference path="../.astro/types.d.ts" />
