+++ apps/docs/src/layouts/MainLayout.astro
@@ -8,9 +8,9 @@ import "@fontsource/lora/500.css";
 import "@fontsource/lora/600.css";
 import "@fontsource/lora/700.css";
 import "@fontsource/jetbrains-mono/400.css";
-import "@bibleio/design/dist/index.css";
+import "@/styles/global.css";
 
-import Footer from "../components/Footer.astro";
+import { Footer } from "@bibleio/design";
 import { Nav } from "../components/Nav";
 
 const { title } = Astro.props;
@@ -28,6 +28,6 @@ const { title } = Astro.props;
   >
     <Nav path={Astro.url.pathname} client:load />
     <slot />
-    <Footer  />
+    <Footer hrefTargetOutside />
   </body>
 </html>
