+++ apps/docs/src/pages/api/docs/[...slug].astro
@@ -1,8 +1,8 @@
 ---
-import { getCollection } from "astro:content";
-import Prose from "@/Prose.astro";
-import DocsLayout from "@layouts/DocsLayout.astro";
-import MainLayout from "@layouts/MainLayout.astro";
+import { getCollection, type CollectionEntry } from "astro:content";
+import { Prose } from "@bibleio/design";
+import DocsLayout from "@/layouts/DocsLayout.astro";
+import MainLayout from "@/layouts/MainLayout.astro";
 
 const apiDocs = await getCollection("api");
