+++ apps/docs/src/pages/biblejson/docs/index.astro
@@ -1,7 +1,7 @@
 ---
 import { getCollection } from "astro:content";
-import DocsLayout from "@layouts/DocsLayout.astro";
-import MainLayout from "@layouts/MainLayout.astro";
+import DocsLayout from "@/layouts/DocsLayout.astro";
+import MainLayout from "@/layouts/MainLayout.astro";
 
 const bibleJsonDocs = await getCollection("biblejson");
 ---
