+++ apps/docs/src/pages/design/docs/[...slug].astro
@@ -1,8 +1,8 @@
 ---
 import { getCollection } from "astro:content";
-import Prose from "@/Prose.astro";
-import DocsLayout from "@layouts/DocsLayout.astro";
-import MainLayout from "@layouts/MainLayout.astro";
+import { Prose } from "@bibleio/design";
+import DocsLayout from "@/layouts/DocsLayout.astro";
+import MainLayout from "@/layouts/MainLayout.astro";
 
 const designDocs = await getCollection("design");
