+++ apps/docs/src/pages/index.astro
@@ -1,6 +1,6 @@
 ---
 import { Button } from "@bibleio/design";
-import MainLayout from "@layouts/MainLayout.astro";
+import MainLayout from "@/layouts/MainLayout.astro";
 ---
 
 <MainLayout title="Home"
