+++ apps/docs/src/styles/global.css
@@ -1,0 +1,1 @@
+@import "@bibleio/design/src/tailwind.css";
