+++ apps/docs/tsconfig.json
@@ -5,8 +5,7 @@
 		"jsxImportSource": "react",
 		"baseUrl": ".",
 		"paths": {
-			"@/*": ["src/components/*"],
-			"@layouts/*": ["src/layouts/*"]
+			"@/*": ["src/*"]
 		}
 	}
 }
