+++ apps/website/src/layouts/Layout.astro
@@ -1,6 +1,5 @@
 ---
 import "@/styles/global.css";
-import "@bibleio/design/src/tailwind.css";
 import "@fontsource/lora/400.css";
 import "@fontsource/lora/500.css";
 import "@fontsource/lora/600.css";
