+++ apps/website/src/pages/blog/[...slug].astro
@@ -1,7 +1,7 @@
 ---
 import { Image } from "astro:assets";
 import { getCollection, render, type CollectionEntry } from "astro:content";
-import Prose from "@/components/Prose.astro";
+import { Prose } from "@bibleio/design";
 import Layout from "@/layouts/Layout.astro";
 
 const authors = await getCollection("team");
