+++ apps/website/src/pages/handbook/[...slug].astro
@@ -2,7 +2,7 @@
 ---
 import { getCollection, render, type CollectionEntry } from "astro:content";
 import Layout from "@/layouts/Layout.astro";
-import Prose from "@/components/Prose.astro";
+import { Prose } from "@bibleio/design";
 import Sidebar from "@/components/handbook/Sidebar.astro";
 
 export async function getStaticPaths() {
